Restrict is_subkey implementation
The previous implementation was too lax, resulting on
is_subkey <primary-key> <index-key> to be true. Ideally, we want
is_subkey to only work with keys in the same "key bucket". A key bucket
is the set of keys that is part of the same hierarchy.

For tables, the key bucket contains all the keys in the table.

For primary and field keys, the key bucket contains all the other
primary and field keys.

For indices and unique indices, the key bucket contains all the other
index and unique index keys, respectively.

This new implementation checks that the two keys live in the same key
bucket.